Refactor: make use of vmdb.Spec in the main application
This involves a change where step expansion is only done at the
beginning. This doesn't currently matter, but in principle it does. We
use attributes of the vmdb.State object as template variables, and if
one step sets a variable, a later step could use it in a template. If
this ever happens, we'll need to change things so that it works again.
Currently nothing does that, and I'm reluctant to write that code
until it's actually needed. YAGNI and all that.